requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
pyjwt[crypto]>=2.8.0
python-dotenv>=1.0.0
pydantic-settings>=2.0.0
sqlmodel>=0.0.14
psycopg2-binary>=2.9.9
httpx>=0.25.0